tag: 3.0.7+beta
- 更新 modVersion 至 3.0.7+beta，并同步更新依赖 ibukigourd 至 0.7.17。
- 替换 VertexConsumerProviderChecker 为 IrisCompat，并合并函数逻辑。
- 修改 maxWidth 方法调用，移除 textRenderer 参数以使用更简洁的调用方式。
- 将 ChatBubbleHandler 的 maxWidth 配置项从 int 改为 float 类型。
- 修正多个渲染模块中的文字宽度获取方式，使用统一的 width 属性。

+++ src/client/kotlin/moe/forpleuvoir/hiirosakura/functional/chataddons/chatbubble/ChatBubbleHandler.kt
@@ -8,7 +8,7 @@ import moe.forpleuvoir.nebula.common.color.Color
 import moe.forpleuvoir.nebula.config.item.impl.boolean
 import moe.forpleuvoir.nebula.config.item.impl.color
 import moe.forpleuvoir.nebula.config.item.impl.duration
-import moe.forpleuvoir.nebula.config.item.impl.int
+import moe.forpleuvoir.nebula.config.item.impl.float
 import net.minecraft.client.render.VertexConsumerProvider
 import net.minecraft.client.util.math.MatrixStack
 import org.joml.Vector2f
@@ -26,7 +26,7 @@ object ChatBubbleHandler:ModConfigContainer("chat_bubble") {
 
     val scale by vector2f("scale", Vector2f(1f, 1f), Vector2f(0.1f, 0.1f), Vector2f(10f, 10f))
 
-    val maxWidth by int("max_width", 120, 9, 480)
+    val maxWidth by float("max_width", 120f, 9f, 480f)
 
     val textureColor by color("texture_color", Color("#FFFF4646"))
